# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Burnout Monitor API
  description: API for employee burnout detection and peak performance identification
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        employeeName:
          type: string
        type:
          type: string
          enum: [burnout, opportunity]
        zone:
          type: string
          enum: [red, green]
        title:
          type: string
        message:
          type: string
        isAcknowledged:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: Get alerts
      description: Manager gets all alerts; employee gets only their own
      parameters:
        - name: acknowledged
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
            enum: [burnout, opportunity]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
  /alerts/{id}/acknowledge:
    put:
      tags: [Alerts]
      summary: Acknowledge an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'

  # ==========================================
  # Notifications
  # ==========================================
...
  /demo/trigger-alert:
    post:
      tags: [Demo]
      summary: Force a zone transition for demo
      description: Manager only - triggers alert for specified employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employeeId
                - targetZone
              properties:
                employeeId:
                  type: string
                  format: uuid
                targetZone:
                  type: string
                  enum: [red, green]
      responses:
        '200':
          description: Alert triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /demo/advance-time:
    post:
      tags: [Demo]
      summary: Advance simulation time
      description: Manager only - fast-forward virtual clock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  description: Number of days to advance
                  default: 1
      responses:
        '200':
          description: Time advanced
          content:
            application/json:
              schema:
                type: object
                properties:
                  virtualTime:
                    type: string
                    format: date-time
                  alertsGenerated:
                    type: integer
...